// frontend/src/api/model/trackingType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Gym Tracker API
 * OpenAPI spec version: 1.0.0
 */

export type TrackingType = typeof TrackingType[keyof typeof TrackingType];


export const TrackingType = {
  reps_weight: 'reps_weight',
  bodyweight: 'bodyweight',
  duration: 'duration',
} as const;